requirements.txt
SimpleITK>=1.2.4
opencv-python
matplotlib
numpy>=1.18.1
tqdm>=4.46.1
torch>=1.15.0
torchvision>=0.6.0
pillow>=7.1.2